python-scripts/requirements.txt
fastapi
uvicorn
selenium
python-dotenv
pymongo 